requirements.txt
aiogram
openai
deepgram-sdk